# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "OGC API Connected Systems implementation (Part 1 + Part 2)",
        "title": "OGC API - Connected Systems",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/controlstreams": {
            "get": {
                "description": "Returns a paginated collection of control stream resources",
                "tags": [
                    "Control Streams"
                ],
                "summary": "List control streams",
                "parameters": [
                    {
                        "description": "Maximum number of results",
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Opaque pagination cursor",
                        "name": "cursor",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated resource IDs",
                        "name": "id",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated keywords for full-text search",
                        "name": "q",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated system IDs",
                        "name": "system",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated feature of interest IDs",
                        "name": "foi",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Comma-separated controlled property IDs",
                        "name": "controlledProperty",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Issue time filter (RFC 3339 date-time or interval)",
                        "name": "issueTime",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Execution time filter (RFC 3339 date-time or interval)",
                        "name": "executionTime",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ControlStreamCollectionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/controlstreams/{controlStreamId}": {
            "get": {
                "description": "Returns a single control stream resource by ID",
                "tags": [
                    "Control Streams"
                ],
                "summary": "Get control stream",
                "parameters": [
                    {
                        "description": "Control Stream ID",
                        "name": "controlStreamId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Replaces a control stream resource by ID",
                "tags": [
                    "Control Streams"
                ],
                "summary": "Update control stream",
                "parameters": [
                    {
                        "description": "Control Stream ID",
                        "name": "controlStreamId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/Controlstream"
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a control stream resource by ID",
                "tags": [
                    "Control Streams"
                ],
                "summary": "Delete control stream",
                "parameters": [
                    {
                        "description": "Control Stream ID",
                        "name": "controlStreamId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Cascade delete to dependent resources",
                        "name": "cascade",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "//localhost:8080"
        }
    ],
    "components": {
        "requestBodies": {
...
            "Controlstream": {
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                },
                "description": "Control Stream resource",
                "required": true
            },
...
        },
        "schemas": {
...
            "api.ControlStreamCollectionResponse": {
                "type": "object",
                "properties": {
                    "items": {
                        "type": "array",
                        "items": {}
                    },
                    "links": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/common_shared.Link"
                        }
                    }
                }
            },
...
            "common_shared.Link": {
                "type": "object",
                "properties": {
                    "href": {
                        "type": "string"
                    },
                    "rel": {
                        "type": "string"
                    },
                    "title": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "uid": {
                        "type": "string"
                    }
                }
            }
        }
    }
}